+++ src/app/app.component.scss
@@ -1,0 +1,6 @@
+.general-container{
+    height: 100vh;
+    background: url('../../src/assets/img/background.jpg') no-repeat center;
+    background-size: cover;
+    padding: 0 10px;
+}
