+++ src/app/components/dashboard/dashboard.component.html
@@ -1,0 +1,75 @@
+<div class="dashboard-container">
+  <div class="dashboard-column">
+    <div class="dashboard-task">
+      <div class="task-name">
+        Test name
+      </div>
+      <div class="task-edit__icon">
+        Edit
+      </div>
+    </div>
+    <div class="dashboard-task">
+      <div class="task-name">
+        Test name
+      </div>
+      <div class="task-edit__icon">
+        Edit
+      </div>
+    </div>
+    <div class="dashboard-task">
+      <div class="task-name">
+        Test name
+      </div>
+      <div class="task-edit__icon">
+        Edit
+      </div>
+    </div>
+  </div>
+  <div class="dashboard-btn-container">
+    <div class="add-task-btn">
+      + Add list
+    </div>
+  </div>
+</div>
+
+
+<!-- <div class="example-container">
+  <h2>To do</h2>
+  <div
+    cdkDropList
+    #todoList="cdkDropList"
+    [cdkDropListData]="todo"
+    [cdkDropListConnectedTo]="[doneList, test1list]"
+    class="example-list"
+    (cdkDropListDropped)="drop($event)">
+    <div class="example-box" *ngFor="let item of todo" cdkDrag>{{item}}</div>
+  </div>
+</div>
+
+<div class="example-container">
+  <h2>Test 1</h2>
+
+  <div
+    cdkDropList
+    #test1list="cdkDropList"
+    [cdkDropListData]="test1"
+    [cdkDropListConnectedTo]="[doneList, todoList]"
+    class="example-list"
+    (cdkDropListDropped)="drop($event)">
+    <div class="example-box" *ngFor="let item of test1" cdkDrag>{{item}}</div>
+  </div>
+</div>
+
+<div class="example-container">
+  <h2>Done</h2>
+
+  <div
+    cdkDropList
+    #doneList="cdkDropList"
+    [cdkDropListData]="done"
+    [cdkDropListConnectedTo]="[todoList, test1list]"
+    class="example-list"
+    (cdkDropListDropped)="drop($event)">
+    <div class="example-box" *ngFor="let item of done" cdkDrag>{{item}}</div>
+  </div>
+</div> -->
