+++ src/app/components/dashboard/dashboard.component.scss
@@ -1,0 +1,102 @@
+@import '../../../variables.scss';
+
+
+.dashboard-container{
+    display: flex;
+}
+
+.dashboard-column{
+    background: $background-grey;
+    min-width: 270px;
+    height: max-content;
+    border-radius: 3px;
+    margin-right: 10px;
+    padding: 8px;
+}
+
+.dashboard-task{
+    padding: 8px;
+    background: white;
+    margin-bottom: 8px;
+
+    &:last-child{
+        margin-bottom: 0;
+    }
+}
+
+.dashboard-btn-container{
+  margin-right: 10px;
+  padding: 0 8px;
+  min-width: 270px;
+  min-height: 32px;
+  height: max-content;
+
+}
+
+.add-task-btn{
+  display: flex;
+  align-items: center;
+  padding: 6px 8px;
+  font-weight: bold;
+  color: white;
+  background: rgba(0, 0, 0, 0.3);
+  border-radius: 3px;
+  cursor: pointer;
+}
+
+
+//////////////
+.example-container {
+    width: 400px;
+    max-width: 100%;
+    margin: 0 25px 25px 0;
+    display: inline-block;
+    vertical-align: top;
+  }
+  
+  .example-list {
+    border: solid 1px #ccc;
+    min-height: 60px;
+    background: white;
+    border-radius: 4px;
+    overflow: hidden;
+    display: block;
+  }
+  
+  .example-box {
+    padding: 20px 10px;
+    border-bottom: solid 1px #ccc;
+    color: rgba(0, 0, 0, 0.87);
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
+    box-sizing: border-box;
+    cursor: move;
+    background: white;
+    font-size: 14px;
+  }
+  
+  .cdk-drag-preview {
+    box-sizing: border-box;
+    border-radius: 4px;
+    box-shadow: 0 5px 5px -3px rgba(0, 0, 0, 0.2),
+                0 8px 10px 1px rgba(0, 0, 0, 0.14),
+                0 3px 14px 2px rgba(0, 0, 0, 0.12);
+  }
+  
+  .cdk-drag-placeholder {
+    opacity: 0;
+  }
+  
+  .cdk-drag-animating {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+  }
+  
+  .example-box:last-child {
+    border: none;
+  }
+  
+  .example-list.cdk-drop-list-dragging .example-box:not(.cdk-drag-placeholder) {
+    transition: transform 250ms cubic-bezier(0, 0, 0.2, 1);
+  }
