+++ src/app/components/navbartop/navbartop.component.html
@@ -1,0 +1,3 @@
+<div class="header-container">
+  navbartop works!
+</div>
