+++ src/app/components/navbartop/navbartop.component.scss
@@ -1,0 +1,7 @@
+.header-container{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 50px;
+    margin-bottom: 10px;
+}
