--- src/app/dashboard/dashboard.component.html
@@ -1,3 +1,0 @@
-<p>
-  dashboard works!
-</p>
