--- src/app/navbartop/navbartop.component.html
@@ -1,3 +1,0 @@
-<p>
-  navbartop works!
-</p>
