+++ src/styles.scss
@@ -1,1 +1,5 @@
 /* You can add global styles to this file, and also import other style files */
+
+body{
+    margin: 0;
+}
