+++ src/variables.scss
@@ -1,0 +1,3 @@
+// Colors
+
+$background-grey: #dfe3e6;
